Moving sdk shared tests
